Moving test_element
