+++ kratos/CMakeLists.txt
@@ -62,7 +62,7 @@ set( KRATOS_CORE_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/modified_shape_functions/triangle_2d_3_ausas_modified_shape_functions.cpp;
     ${CMAKE_CURRENT_SOURCE_DIR}/modified_shape_functions/tetrahedra_3d_4_modified_shape_functions.cpp;
     ${CMAKE_CURRENT_SOURCE_DIR}/modified_shape_functions/tetrahedra_3d_4_ausas_modified_shape_functions.cpp;
-    ${CMAKE_CURRENT_SOURCE_DIR}/elements/test_element.cpp;
+    ${CMAKE_CURRENT_SOURCE_DIR}/tests/test_element.cpp;
 )
 
 ## Kratos python interface code
